[REFACTOR] repository directory 수정
