+++ user-api/src/main/java/com/biengual/userapi/user/domain/UserCustomRepository.java
@@ -1,7 +1,5 @@
-package com.biengual.userapi.user.repository;
+package com.biengual.userapi.user.domain;
 
-import com.biengual.userapi.user.domain.QUserEntity;
-import com.biengual.userapi.user.domain.UserInfo;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
